[Refactoring ⚙️] 폴더 내 파일 묶음
