+++ frontend/src/pages/PlaceDetailPage/PlaceDetail.tsx
@@ -7,9 +7,9 @@ import { useLocation, useParams } from 'react-router-dom'
 
 import AccomodationList from './components/AccomodationList'
 import ChatDrawer from './components/ChatDrawer/ChatDrawer'
-import ContentType12 from './components/ContentType12'
-import ContentType14 from './components/ContentType14'
-import ContentType15 from './components/ContentType15'
+import ContentType12 from './components/ContentType/ContentType12'
+import ContentType14 from './components/ContentType/ContentType14'
+import ContentType15 from './components/ContentType/ContentType15'
 import MiddleMenuBar from './components/MiddleMenuBar'
 import PlaceMap from './components/PlaceMap'
 import * as L from './styles/PlaceDetail.style'
