+++ frontend/src/pages/PlaceDetailPage/components/ContentType/ContentType12.tsx
